requirements.txt
matplotlib~=3.9.2
numpy~=1.26.4
sympy~=1.13.3
tabulate~=0.9.0
streamlit~=1.45.0
seaborn~=0.13.2
scikit-learn~=1.6.1
opencv-python~=4.11.0.86
scipy~=1.13.1
config~=0.5.1
pendulum~=3.0.0
integrator~=0.0.dev0
utils~=1.0.2
yfinance~=0.2.55
pandas~=2.2.2
shap~=0.47.1
plotly~=6.0.1
PyAudio~=0.2.14
imageio~=2.37.0
mediapipe~=0.10.21
pytesseract~=0.3.13
pymatgen~=2025.3.10
matminer~=0.9.3
xgboost~=3.0.0
modal~=0.73.157
joblib~=1.4.2
prophet~=1.1.6
cffi~=1.17.1
sounddevice~=0.5.1
streamlit
numpy
matplotlib

DateTime~=5.5
MetaTrader5~=5.0.4993
pydub~=0.25.1
requests~=2.32.3
openai~=1.78.0

streamlit
pandas
numpy
matplotlib
sympy
openai

python-dotenv~=1.1.0
av~=14.3.0
Pygments~=2.19.1
Pint~=0.24.4
wordcloud~=1.9.4
docx~=0.2.4
fpdf~=1.7.2
gensim~=4.3.3
jax~=0.5.3
pillow~=10.4.0
ultralytics~=8.3.145